Working changes
With these changes it works.  It's not the most elegant solution, but it
will work, and there are two ways for it to work.  Originally I had it
sending a Player object to the server, which would get sent out to each
client.  But then I created a GameEvent class with the new x and new y
to change to.  That way it's more extensible and it's not sending as big
of a data type.  Both ways are still accepted, however.